refactored worker architecture
