+++ .gitignore
@@ -1,5 +1,6 @@
 
 input_data
+output-*
 
 *.class
