+++ src/main/java/ch/sbb/matsim/umlego/WorkItem.java
@@ -1,0 +1,9 @@
+package ch.sbb.matsim.umlego;
+
+import java.util.concurrent.CompletableFuture;
+
+public record WorkItem(
+        String originZone,
+        CompletableFuture<WorkResult> result
+) {
+}
